+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added `classifications` field in `Family` type
+
 ## [0.7.7] - 2025-07-11
 
 - Added `PricingRule` class with properties: tier (int), type (string), rateType (string), and value (float)